Fix return types for _smu_get_hw_version, _psp_sw_init

+++ WhateverRed/kern_wred.hpp
@@ -106,8 +106,8 @@ class WRed {
 
     /** X5000HWLibs */
     static void wrapAmdTtlServicesConstructor(void *that, IOPCIDevice *provider);
-    static uint64_t wrapSmuGetHwVersion(uint64_t param1, uint32_t param2);
-    static uint64_t wrapPspSwInit(uint32_t *param1, uint32_t *param2);
+    static uint32_t wrapSmuGetHwVersion(uint64_t param1, uint32_t param2);
+    static uint32_t wrapPspSwInit(uint32_t *param1, uint32_t *param2);
     static uint32_t wrapGcGetHwVersion(uint32_t *param1);
     static void wrapPopulateFirmwareDirectory(void *that);
     static void *wrapCreatePowerTuneServices(void *that, void *param2);
